Add support for sdkVersion in client registration
