+++ lib/db/client-instance-store.js
@@ -5,6 +5,7 @@ const logger = require('../logger');
 const COLUMNS = [
     'app_name',
     'instance_id',
+    'sdk_version',
     'client_ip',
     'last_seen',
     'created_at',
@@ -14,6 +15,7 @@ const TABLE = 'client_instances';
 const mapRow = row => ({
     appName: row.app_name,
     instanceId: row.instance_id,
+    sdkVersion: row.sdk_version,
     clientIp: row.client_ip,
     lastSeen: row.last_seen,
     createdAt: row.created_at,
@@ -48,6 +50,7 @@ class ClientInstanceStore {
             .update({
                 last_seen: 'now()',
                 client_ip: details.clientIp,
+                sdk_version: details.sdkVersion,
             });
     }
 
@@ -55,6 +58,7 @@ class ClientInstanceStore {
         return this.db(TABLE).insert({
             app_name: details.appName,
             instance_id: details.instanceId,
+            sdk_version: details.sdkVersion,
             client_ip: details.clientIp,
         });
     }
