+++ lib/routes/client-api/metrics-schema.js
@@ -12,12 +12,4 @@ const clientMetricsSchema = joi.object().keys({
     }),
 });
 
-const clientRegisterSchema = joi.object().keys({
-    appName: joi.string().required(),
-    instanceId: joi.string().required(),
-    strategies: joi.array().required().items(joi.string(), joi.any().strip()),
-    started: joi.date().required(),
-    interval: joi.number().required(),
-});
-
-module.exports = { clientMetricsSchema, clientRegisterSchema };
+module.exports = { clientMetricsSchema };
