+++ lib/routes/client-api/register.js
@@ -4,7 +4,7 @@ const { Router } = require('express');
 const joi = require('joi');
 const logger = require('../../logger');
 
-const { clientRegisterSchema } = require('./metrics-schema');
+const { clientRegisterSchema } = require('./register-schema');
 
 exports.router = config => {
     const { clientInstanceStore, clientApplicationsStore } = config.stores;
