+++ lib/routes/client-api/register.test.js
@@ -36,6 +36,22 @@ test('should register client', t => {
             appName: 'demo',
             instanceId: 'test',
             strategies: ['default'],
+            sdkVersion: 'unleash-client-test:1.2',
+            started: Date.now(),
+            interval: 10,
+        })
+        .expect(202);
+});
+
+test('should register client without sdkVersin', t => {
+    t.plan(0);
+    const { request } = getSetup();
+    return request
+        .post('/api/client/register')
+        .send({
+            appName: 'demo',
+            instanceId: 'test',
+            strategies: ['default'],
             started: Date.now(),
             interval: 10,
         })
